requirements.txt
# API / backend
fastapi==0.115.*
uvicorn[standard]==0.30.*
pydantic==2.*
python-dotenv==1.0.*
httpx==0.27.*
tenacity==8.5.*

# UI
streamlit==1.37.*
